Optionally bracket all library with namespace
The ESP8266 (and probably others) have conflicting class naes for "File"
and a few other names defined in SdFat.

To guard against this, when the symbol "SDFAT_NAMESPACE" is defined at
compile-time, bracket all objects, globals, constants, etc. in a
"namespace sdfat { ... };" block.

All examples updated to add "using namespace sdfat;" when required,
meaning that they compile and run whether or not the define is set.

No changes to existing apps are needed.