Remove `undef f` from SysCall.h
When SdFat.h is included it was undefining the F() (flash string) macro.
This causes all strings to go into RAM, as seen in this issue:
https://github.com/esp8266/Arduino/issues/6767

Remove this undef.  The F() is only used in a few print statements which
are F()-safe on the 8266 now, and in the SD.h and SDFS.h libs the
function that actually calls print(F(xxxx)) is never used.